Allow report generators to skip output for pages
This commit allows the report generators to skip sending output for certain pages, if there isn't any relevant output to record in the CSV for a particular report. This saves us from generating reports that might contain many empty rows, which isn't useful for anyone.

Report generators can take advantage of this by returning an empty array as the processing result for a particular GOV.UK page, which is then checked by the `ReportRunner` before being pushed on to the CSV write queue (this is more efficient than having the CSV writer check, as the CSV writer will already be busy and any more processing we push to it will only slow things down).