Add link text report
This commit adds a link text report, which checks for GOV.UK pages containing links with the same text that point to different URLs. This report was originally written by Oscar in the Knowledge Graph repo, and this commit ports it over to be run via the `ReportRunner`, whilst refactoring to remove all logic requesting GOV.UK pages and instead using the information we have in the content item and HTML markup for any particular page. It also includes some additional filters for links that the Accessibility team don't wish to appear in the results (such as requesting accessible versions of documents etc).

Co-Authored-By: Oscar Wyatt <oscar.wyatt@digital.cabinet-office.gov.uk>