Check for existence of href in <a> tags
This commit checks for the existence of the `href` property of `<a>` tags before trying to get the value. From running the link text report generator this problem was noticed - we might want to call this out separately in the future. If the `href` property isn't found for a link, we set the link to `'https://www.gov.uk/'` (which we skip over by default).