Add function to extract file extension
This is so we can filter for pages with extensions only.

Also fix other functions that return a list when the TRYCATCH is successful but returns a string otherwise. Are returning a list in both instances so we have a column of lists which will make it easier to filter from.